backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
pyyaml
requests
pydantic-settings
cryptography